agentflow/core/isa/pattern_mining.py
"""Advanced pattern mining implementations."""
from typing import Dict, List, Optional, Set, Tuple, Any
import numpy as np
from dataclasses import dataclass
from collections import defaultdict
from .patterns import Pattern, PatternType, PatternMetrics
from .formal import FormalInstruction
from .analyzer import AnalysisResult

@dataclass
class FrequentSequence:
    """Represents a frequent instruction sequence."""
    instructions: List[FormalInstruction]
    support: float
    confidence: float
    frequency: int

@dataclass
class DependencyGraph:
    """Represents instruction dependencies."""
    nodes: Dict[str, FormalInstruction]
    edges: List[Tuple[str, str]]
    weights: Dict[Tuple[str, str], float]

class SequenceMiner:
    """Advanced sequence pattern mining."""
    
    def __init__(self, config: Dict[str, Any]):
        self.config = config
        self.min_support = config.get("min_support", 0.1)
        self.min_confidence = config.get("min_confidence", 0.8)
        
    def find_sequences(self,
                      instructions: List[FormalInstruction]) -> List[FrequentSequence]:
        """Find frequent instruction sequences."""
        sequences = []
        n = len(instructions)
        
        # Generate candidate sequences
        for length in range(2, min(10, n + 1)):
            candidates = self._generate_candidates(instructions, length)
            
            # Calculate support for each candidate
            for candidate in candidates:
                support = self._calculate_support(candidate, instructions)
                if support >= self.min_support:
                    confidence = self._calculate_confidence(
                        candidate,
                        instructions
                    )
                    if confidence >= self.min_confidence:
                        sequences.append(
                            FrequentSequence(
                                instructions=candidate,
                                support=support,
                                confidence=confidence,
                                frequency=self._count_frequency(
                                    candidate,
                                    instructions
                                )
                            )
                        )
                        
        return sequences
    
    def _generate_candidates(self,
                           instructions: List[FormalInstruction],
                           length: int) -> List[List[FormalInstruction]]:
        """Generate candidate sequences."""
        candidates = []
        n = len(instructions)
        
        for i in range(n - length + 1):
            candidate = instructions[i:i+length]
            if self._is_valid_candidate(candidate):
                candidates.append(candidate)
                
        return candidates
    
    def _is_valid_candidate(self,
                          candidate: List[FormalInstruction]) -> bool:
        """Check if candidate sequence is valid."""
        # Check for dependencies
        for i in range(len(candidate) - 1):
            if self._has_dependency(candidate[i], candidate[i+1]):
                return True
        return False
    
    def _has_dependency(self,
                       instr1: FormalInstruction,
                       instr2: FormalInstruction) -> bool:
        """Check if instructions have dependency."""
        # Implement dependency check
        return True  # Placeholder

class ParallelMiner:
    """Advanced parallel pattern mining."""
    
    def __init__(self, config: Dict[str, Any]):
        self.config = config
        
    def find_parallel_groups(self,
                           instructions: List[FormalInstruction]
                           ) -> List[List[FormalInstruction]]:
        """Find groups of instructions that can be executed in parallel."""
        # Build dependency graph
        graph = self._build_dependency_graph(instructions)
        
        # Find independent sets of instructions
        independent_sets = self._find_independent_sets(graph)
        
        # Convert node IDs back to instructions
        parallel_groups = []
        for node_set in independent_sets:
            group = [graph.nodes[node] for node in node_set]
            parallel_groups.append(group)
            
        return parallel_groups
    
    def _build_dependency_graph(self,
                              instructions: List[FormalInstruction]
                              ) -> DependencyGraph:
        """Build instruction dependency graph."""
        nodes = {str(i): instr for i, instr in enumerate(instructions)}
        edges = []
        weights = {}
        
        # Add edges for dependencies
        for i, instr1 in enumerate(instructions):
            for j, instr2 in enumerate(instructions[i+1:], i+1):
                if self._has_dependency(instr1, instr2):
                    edge = (str(i), str(j))
                    edges.append(edge)
                    weights[edge] = 1.0  # Default weight
                    
        return DependencyGraph(
            nodes=nodes,
            edges=edges,
            weights=weights
        )
        
    def _find_independent_sets(self,
                             graph: DependencyGraph
                             ) -> List[Set[str]]:
        """Find sets of independent instructions."""
        independent_sets = []
        remaining_nodes = set(graph.nodes.keys())
        
        while remaining_nodes:
            # Find maximal independent set
            independent_set = self._find_maximal_independent_set(
                remaining_nodes,
                graph
            )
            
            if not independent_set:
                break
                
            independent_sets.append(independent_set)
            remaining_nodes -= independent_set
            
        return independent_sets
    
    def _find_maximal_independent_set(self,
                                    nodes: Set[str],
                                    graph: DependencyGraph
                                    ) -> Set[str]:
        """Find maximal independent set in graph."""
        independent_set = set()
        candidates = nodes.copy()
        
        while candidates:
            # Choose node with minimum degree
            min_degree_node = min(
                candidates,
                key=lambda n: self._calculate_node_degree(n, graph)
            )
            
            independent_set.add(min_degree_node)
            candidates.remove(min_degree_node)
            
            # Remove neighbors of chosen node
            neighbors = self._get_neighbors(min_degree_node, graph)
            candidates -= neighbors
            
        return independent_set
    
    def _calculate_node_degree(self,
                             node: str,
                             graph: DependencyGraph
                             ) -> int:
        """Calculate degree of node in graph."""
        return len(self._get_neighbors(node, graph))
    
    def _get_neighbors(self,
                      node: str,
                      graph: DependencyGraph
                      ) -> Set[str]:
        """Get neighbors of node in graph."""
        neighbors = set()
        for edge in graph.edges:
            if node == edge[0]:
                neighbors.add(edge[1])
            elif node == edge[1]:
                neighbors.add(edge[0])
        return neighbors
        
    def _has_dependency(self,
                       instr1: FormalInstruction,
                       instr2: FormalInstruction
                       ) -> bool:
        """Check if instructions have dependency."""
        # Check for data dependencies
        if any(
            out in instr2.get_inputs()
            for out in instr1.get_outputs()
        ):
            return True
            
        # Check for resource dependencies
        if any(
            res1 == res2
            for res1 in instr1.get_resources()
            for res2 in instr2.get_resources()
        ):
            return True
            
        # Check for ordering dependencies
        if instr1.requires_ordering() or instr2.requires_ordering():
            return True
            
        return False

class BehavioralMiner:
    """Advanced behavioral pattern mining."""
    
    def __init__(self, config: Dict[str, Any]):
        self.config = config
        self.min_pattern_length = config.get("min_pattern_length", 2)
        self.similarity_threshold = config.get("similarity_threshold", 0.8)
        
    def find_behavioral_patterns(self,
                               instructions: List[FormalInstruction],
                               analysis: AnalysisResult
                               ) -> List[Pattern]:
        """Find behavioral patterns in instructions."""
        patterns = []
        
        # Extract behavioral features
        features = self._extract_behavioral_features(
            instructions,
            analysis
        )
        
        # Cluster similar behaviors
        clusters = self._cluster_behaviors(features)
        
        # Extract patterns from clusters
        for cluster in clusters:
            pattern = self._extract_pattern_from_cluster(
                cluster,
                instructions,
                analysis
            )
            if pattern:
                patterns.append(pattern)
                
        return patterns
    
    def find_patterns(self,
                     instructions: List[FormalInstruction]
                     ) -> List[List[FormalInstruction]]:
        """Find behavioral patterns in instructions."""
        if not instructions:
            return []
            
        patterns = []
        n = len(instructions)
        
        # Find patterns of different lengths
        for length in range(self.min_pattern_length, min(n + 1, 10)):
            for i in range(n - length + 1):
                candidate = instructions[i:i+length]
                
                # Check if candidate is a valid pattern
                if self._is_valid_pattern(candidate):
                    patterns.append(candidate)
                    
        return patterns
        
    def calculate_pattern_similarity(self,
                                   pattern1: List[FormalInstruction],
                                   pattern2: List[FormalInstruction]
                                   ) -> float:
        """Calculate similarity between two patterns."""
        if not pattern1 or not pattern2:
            return 0.0
            
        # Compare instruction sequences
        similarity_scores = []
        
        # Compare lengths
        len_similarity = min(len(pattern1), len(pattern2)) / max(len(pattern1), len(pattern2))
        similarity_scores.append(len_similarity)
        
        # Compare instruction names
        name_matches = sum(
            1 for i1, i2 in zip(pattern1, pattern2)
            if i1.name == i2.name
        )
        name_similarity = name_matches / max(len(pattern1), len(pattern2))
        similarity_scores.append(name_similarity)
        
        # Compare parameters
        param_similarities = []
        for i1, i2 in zip(pattern1, pattern2):
            if i1.name == i2.name:
                param_match = sum(
                    1 for k, v in i1.params.items()
                    if k in i2.params and i2.params[k] == v
                )
                param_total = len(set(i1.params) | set(i2.params))
                param_similarities.append(
                    param_match / param_total if param_total > 0 else 1.0
                )
                
        if param_similarities:
            param_similarity = sum(param_similarities) / len(param_similarities)
            similarity_scores.append(param_similarity)
            
        return sum(similarity_scores) / len(similarity_scores)
        
    def calculate_pattern_frequency(self,
                                  pattern: List[FormalInstruction],
                                  instructions: List[FormalInstruction]
                                  ) -> int:
        """Calculate frequency of pattern in instruction sequence."""
        if not pattern or not instructions:
            return 0
            
        frequency = 0
        pattern_len = len(pattern)
        
        # Slide window over instructions
        for i in range(len(instructions) - pattern_len + 1):
            window = instructions[i:i+pattern_len]
            
            # Check if window matches pattern
            if self.calculate_pattern_similarity(pattern, window) >= self.similarity_threshold:
                frequency += 1
                
        return frequency
        
    def optimize_sequence(self,
                        instructions: List[FormalInstruction]
                        ) -> List[FormalInstruction]:
        """Optimize instruction sequence using patterns."""
        if not instructions:
            return []
            
        # Find patterns
        patterns = self.find_patterns(instructions)
        if not patterns:
            return instructions.copy()
            
        # Score patterns
        pattern_scores = [
            (pattern, self._calculate_pattern_score(pattern))
            for pattern in patterns
        ]
        
        # Sort by score
        pattern_scores.sort(key=lambda x: x[1], reverse=True)
        
        # Build optimized sequence
        optimized = []
        used = set()
        
        # Add highest scoring patterns first
        for pattern, score in pattern_scores:
            for i in range(len(instructions) - len(pattern) + 1):
                window = instructions[i:i+len(pattern)]
                if (self.calculate_pattern_similarity(pattern, window) >= self.similarity_threshold and
                    not any(j in used for j in range(i, i+len(pattern)))):
                    optimized.extend(window)
                    used.update(range(i, i+len(pattern)))
                    
        # Add remaining instructions
        for i, instr in enumerate(instructions):
            if i not in used:
                optimized.append(instr)
                
        return optimized
        
    def _is_valid_pattern(self,
                         pattern: List[FormalInstruction]
                         ) -> bool:
        """Check if pattern is valid."""
        if len(pattern) < self.min_pattern_length:
            return False
            
        # Check for meaningful relationships
        for i in range(len(pattern) - 1):
            if not self._has_relationship(pattern[i], pattern[i+1]):
                return False
                
        return True
        
    def _has_relationship(self,
                         instr1: FormalInstruction,
                         instr2: FormalInstruction
                         ) -> bool:
        """Check if instructions have meaningful relationship."""
        # Check for data flow
        if any(
            out in instr2.get_inputs()
            for out in instr1.get_outputs()
        ):
            return True
            
        # Check for control flow
        if instr1.name in ["if", "while", "for"] and instr2.name in ["endif", "endwhile", "endfor"]:
            return True
            
        # Check for semantic relationship
        if instr1.name in ["load", "read"] and instr2.name in ["process", "validate"]:
            return True
        if instr1.name in ["process", "validate"] and instr2.name in ["save", "write"]:
            return True
            
        return False
        
    def _calculate_pattern_score(self,
                               pattern: List[FormalInstruction]
                               ) -> float:
        """Calculate score for pattern."""
        # Consider factors like:
        # - Pattern length
        length_score = min(len(pattern) / 5.0, 1.0)
        
        # - Instruction relationships
        relationship_score = sum(
            1 for i in range(len(pattern)-1)
            if self._has_relationship(pattern[i], pattern[i+1])
        ) / (len(pattern) - 1)
        
        # - Pattern cohesion
        cohesion_score = self._calculate_cohesion(pattern)
        
        return (length_score + relationship_score + cohesion_score) / 3.0
        
    def _calculate_cohesion(self,
                          pattern: List[FormalInstruction]
                          ) -> float:
        """Calculate pattern cohesion."""
        if len(pattern) < 2:
            return 0.0
            
        # Calculate shared resources
        resources = defaultdict(set)
        for i, instr in enumerate(pattern):
            for res in instr.get_resources():
                resources[res].add(i)
                
        # Calculate cohesion based on resource sharing
        total_pairs = (len(pattern) * (len(pattern) - 1)) / 2
        shared_pairs = sum(
            len(positions) * (len(positions) - 1) / 2
            for positions in resources.values()
        )
        
        return shared_pairs / total_pairs if total_pairs > 0 else 0.0
        
    def _extract_behavioral_features(self,
                                  instructions: List[FormalInstruction],
                                  analysis: AnalysisResult
                                  ) -> Dict[str, np.ndarray]:
        """Extract behavioral features from instructions."""
        features = {}
        
        for i, instr in enumerate(instructions):
            # Extract features like execution time, resource usage, dependencies
            feature_vector = np.array([
                float(analysis.metrics.get(f"exec_time_{i}", 0.0)),
                float(analysis.metrics.get(f"resource_usage_{i}", 0.0)),
                float(analysis.metrics.get(f"dependency_count_{i}", 0.0)),
                float(analysis.metrics.get(f"complexity_{i}", 0.0))
            ])
            features[str(i)] = feature_vector
            
        return features
    
    def _cluster_behaviors(self,
                         features: Dict[str, np.ndarray]
                         ) -> List[List[str]]:
        """Cluster instructions based on behavioral features."""
        clusters = []
        unassigned = set(features.keys())
        
        while unassigned:
            # Pick a random unassigned instruction as cluster center
            center = next(iter(unassigned))
            cluster = [center]
            unassigned.remove(center)
            
            # Find similar instructions
            for instr_id in list(unassigned):
                if self._calculate_similarity(
                    features[center],
                    features[instr_id]
                ) >= self.similarity_threshold:
                    cluster.append(instr_id)
                    unassigned.remove(instr_id)
                    
            clusters.append(cluster)
            
        return clusters
    
    def _extract_pattern_from_cluster(self,
                                    cluster: List[str],
                                    instructions: List[FormalInstruction],
                                    analysis: AnalysisResult
                                    ) -> Optional[Pattern]:
        """Extract pattern from instruction cluster."""
        if len(cluster) < self.min_pattern_length:
            return None
            
        # Get instructions in cluster
        cluster_instructions = [
            instructions[int(i)] for i in cluster
        ]
        
        # Calculate pattern metrics
        metrics = self._calculate_pattern_metrics(
            cluster_instructions,
            analysis
        )
        
        return Pattern(
            type=PatternType.BEHAVIORAL,
            instructions=cluster_instructions,
            metrics=metrics
        )
    
    def _calculate_similarity(self,
                            feature1: np.ndarray,
                            feature2: np.ndarray
                            ) -> float:
        """Calculate similarity between feature vectors."""
        return 1.0 / (1.0 + np.linalg.norm(feature1 - feature2))
        
    def _calculate_pattern_metrics(self,
                                instructions: List[FormalInstruction],
                                analysis: AnalysisResult
                                ) -> PatternMetrics:
        """Calculate metrics for behavioral pattern."""
        # Calculate average metrics across instructions
        avg_exec_time = np.mean([
            analysis.metrics.get(f"exec_time_{i}", 0.0)
            for i in range(len(instructions))
        ])
        avg_resource_usage = np.mean([
            analysis.metrics.get(f"resource_usage_{i}", 0.0)
            for i in range(len(instructions))
        ])
        
        return PatternMetrics(
            frequency=1,  # This would be calculated from historical data
            confidence=analysis.confidence,
            support=1.0,  # This would be calculated from historical data
            performance_impact=avg_exec_time,
            resource_impact=avg_resource_usage
        )
